backend/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.4
django-cors-headers==4.6.0
django-import-export==4.3.3
django-rest==0.8.7
djangorestframework==3.15.2
djangorestframework-simplejwt==4.7.2
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.3
tablib==3.7.0
djoser==2.1.0
django-extensions==3.2.1
drf-extra-fields==3.7.0
django-filter==23.1